% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LinReg1.R
\name{LinReg1}
\alias{LinReg1}
\title{Linear Regression, One Group Time Trend}
\usage{
LinReg1(data)
}
\arguments{
\item{data}{a m*n*t.max array of gene expression data, where m is number of genes, n is number of subjects at each time point, and t.max is number of time points}
}
\value{
vector of t-statistics for slope coefficient
}
\description{
For one group, time trend simple linear regression model of gene expression. Function fits OLS linear regression model to each gene. Returns t-statistic for test of zero slope.
}

