% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LinReg2.R
\name{LinReg2}
\alias{LinReg2}
\title{Linear Regression, Two Groups Time Trend}
\usage{
LinReg2(data, I.group)
}
\arguments{
\item{data}{a m*n*t.max array of gene expression data, where m is number of genes, n is total number of subjects at each time point (i.e. the sum of the number of subjects in each group, n = n1 + n2, assumes n1 = n2), and t.max is number of time points}

\item{I.group}{vector of 0/1 variables indicating group membership for each subject at each time point, length n*t.max. For example, I.group = rep(c(rep(0, n1), rep(1, n2)), t.max).}
}
\value{
vector of t-statistics for interaction term coefficient
}
\description{
For two group, time trend simple linear regression model of gene expression. Function fits OLS linear regression model to each gene (j = 1, ..., m):
Y.j = beta.0.j + beta.1.j(t) + beta.2.j(I.group) + beta.3.j(t*I.group)
Returns t-statistic for test of different slopes. Assumes group sizes are equal, i.e. n1 = n2.
}

