% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Permutations.R
\name{Permutations}
\alias{Permutations}
\title{Permutations}
\usage{
Permutations(groups, n, t.max, B = 200)
}
\arguments{
\item{n}{sample size per group}

\item{t.max}{number of time points}

\item{B}{number of permutations per significance test, maximum 2000. Recommend 200 for two group single time point, 100 for two group time course, 1000 for one group time course.}

\item{groups:}{number of groups (1 or 2)}
}
\value{
matrix of permutations
}
\description{
Function to generate random permutations. For two groups, permutes labels independently at each time point. For one group time trend, permutes time points.
}

