% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Qval.R
\name{Qval}
\alias{Qval}
\title{Q-value function}
\usage{
Qval(p.values, lambda = 0.5)
}
\arguments{
\item{p.values}{vector of p-values}

\item{lambda}{pFDR tuning parameter (between 0 and 1)}
}
\value{
vector of q-values
}
\description{
Applies positive FDR algorithm to obtain q-value for each gene.
}

