% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimulateData1.R
\name{SimulateData1}
\alias{SimulateData1}
\title{Simulate Data, One Group Time Trend}
\usage{
SimulateData1(n, t.max, pi1, FC, variance, m)
}
\arguments{
\item{n}{number of subjects at each time point}

\item{t.max}{number of time points}

\item{pi1}{proportion differentially expressed genes}

\item{FC}{fold change at time t.max}

\item{variance}{gene expression variance at each time point}

\item{m}{number of genes}
}
\value{
one m*n*t.max array of gene expression data
}
\description{
Generates data for the one group time trend simple linear regression model.
}

