% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimulateData2.R
\name{SimulateData2}
\alias{SimulateData2}
\title{Simulate Data, Two Groups}
\usage{
SimulateData2(n, t.max, pi1, FC, variance, m)
}
\arguments{
\item{n}{number of subjects per group at each time point}

\item{t.max}{number of time points}

\item{pi1}{proportion differentially expressed genes}

\item{FC}{fold change at time t.max}

\item{variance}{gene expression variance at each time point}

\item{m}{number of genes}
}
\value{
one m*n*t.max array of gene expression data
}
\description{
Generates data for two groups single time point (t.max = 1) or the two group time trend simple linear regression model.
}

