% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TestPower.R
\name{TestPower}
\alias{TestPower}
\title{Significance test and power calculation}
\usage{
TestPower(q.values, p.values, f, pi1)
}
\arguments{
\item{q.values}{vector of q-values sorted such that all DE genes come first}

\item{p.values}{vector of corresponding p-values in the same order as the q.values}

\item{f}{desired false discovery rate (between 0 and 1)}

\item{pi1}{the proportion of genes that are truly differentially expressed}
}
\value{
vector of s, r, power, estimated FDR, and true FDR
}
\description{
For simulated microarray data with proportion pi1 of differentially expressed (DE) genes, this function takes a vector of q-values and a vector of the corresponding p-values, sorted such that all differentially expressed genes come first, and returns s (number of true positives), r (number of genes called significant), power (proportion of DE genes that are called significant), estimated false discovery rate (FDR) (maximum q-value <= desired FDR), and true FDR ((r - s)/r if r not equal to zero, otherwise zero).
}

