% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Tstats.R
\name{Tstats}
\alias{Tstats}
\title{T-statistics}
\usage{
Tstats(group.1, group.2)
}
\arguments{
\item{group.1}{group 1 gene expression matrix, rows are genes, columns are subjects}

\item{group.2}{group 2 gene expression matrix, rows are genes, columns are subjects}
}
\value{
vector of t-statistics
}
\description{
Function to calculate t-statistics to test difference of two group means.
}

